renaming phoenicis-script in phoenicis-bash
