+++ phoenicis-bash/pom.xml
@@ -27,7 +27,7 @@
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>phoenicis-scripts</artifactId>
+    <artifactId>phoenicis-bash</artifactId>
 
     <dependencies>
         <dependency>
