+++ phoenicis-bash/src/main/bash/lib/config.sh
