+++ phoenicis-bash/src/main/bash/lib/debug.sh
