+++ phoenicis-bash/src/main/bash/lib/legacy.sh
