+++ phoenicis-bash/src/main/bash/lib/netcat.sh
