+++ phoenicis-bash/src/main/bash/lib/scripts.sh
