+++ phoenicis-bash/src/main/bash/lib/setupWindow.sh
