+++ phoenicis-bash/src/main/bash/lib/shortcut.sh
