+++ phoenicis-bash/src/main/bash/lib/system.sh
