+++ phoenicis-bash/src/main/bash/lib/utils.sh
