+++ phoenicis-bash/src/main/bash/lib/variables.sh
