+++ phoenicis-bash/src/main/bash/lib/wine.sh
