+++ phoenicis-bash/src/main/java/com/playonlinux/bash/ScriptLegacy.java
@@ -16,7 +16,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package com.playonlinux.scripts;
+package com.playonlinux.bash;
 
 import java.io.BufferedReader;
 import java.io.File;
