+++ phoenicis-bash/src/main/java/com/playonlinux/bash/ScriptLegacyFactory.java
@@ -16,7 +16,7 @@
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
 
-package com.playonlinux.scripts;
+package com.playonlinux.bash;
 
 import java.util.concurrent.ExecutorService;
