+++ phoenicis-bash/src/main/python/BashBinder/BashProcess.py
