+++ phoenicis-bash/src/main/python/Environment/EnvironmentFormatter.py
