+++ phoenicis-bash/src/main/python/Environment/EnvironmentLoader.py
