+++ phoenicis-bash/src/main/python/Environment/__init__.py
