+++ phoenicis-bash/src/main/python/SetupWindow/SetupWindowManager.py
