+++ phoenicis-bash/src/test/resources/com/playonlinux/bash/legacyScriptExample.sh
