+++ phoenicis-bash/src/test/resources/com/playonlinux/bash/legacyScriptExampleCRLF.sh
