+++ phoenicis-bash/src/test/resources/com/playonlinux/bash/legacyScriptExampleWithPlayOnLinuxBashHeader.sh
