+++ phoenicis-bash/src/test/resources/com/playonlinux/bash/legacyScriptExampleWithSignature.sh
