+++ phoenicis-bash/src/test/resources/com/playonlinux/bash/legacyScriptExampleWithSignatureCRLF.sh
