+++ phoenicis-bash/src/test/resources/com/playonlinux/bash/realScript.sh
