+++ phoenicis-bash/src/test/resources/com/playonlinux/bash/wrapperTestScript.sh
